+++ src/assetbundles/SettingsAsset.php
@@ -1,0 +1,39 @@
+<?php
+/**
+ * Email Validator plugin for Craft CMS 3.x
+ *
+ * Email address validation for user registrations, custom forms and more.
+ *
+ * @link      https://github.com/lukeyouell
+ * @copyright Copyright (c) 2018 Luke Youell
+ */
+
+namespace lukeyouell\emailvalidator\assetbundles;
+
+use Craft;
+use craft\web\AssetBundle;
+use craft\web\assets\cp\CpAsset;
+
+class SettingsAsset extends AssetBundle
+{
+    // Public Methods
+    // =========================================================================
+
+    /**
+     * @inheritdoc
+     */
+    public function init()
+    {
+        $this->sourcePath = "@lukeyouell/emailvalidator/assetbundles";
+
+        $this->depends = [
+            CpAsset::class,
+        ];
+
+        $this->js = [
+            'js/settings.js',
+        ];
+
+        parent::init();
+    }
+}