+++ src/services/ProviderService.php
@@ -12,7 +12,6 @@ namespace lukeyouell\emailvalidator\services;
 
 use lukeyouell\emailvalidator\EmailValidator;
 use lukeyouell\emailvalidator\models\Provider as ProviderModel;
-use lukeyouell\emailvalidator\records\Provider as ProviderRecord;
 
 use Craft;
 use craft\base\Component;
@@ -20,7 +19,7 @@ use craft\db\Query;
 
 use yii\base\Exception;
 
-class EmailProviderService extends Component
+class ProviderService extends Component
 {
     // Public Methods
     // =========================================================================
@@ -38,6 +37,15 @@ class EmailProviderService extends Component
         }
 
         return $providers;
+    }
+
+    public function countProvidersByType($type)
+    {
+        $count = $this->_createProviderQuery()
+            ->where(['type' => $type])
+            ->count();
+
+        return (int)$count;
     }
 
     public function getProviderByDomain($domain)
@@ -59,38 +67,6 @@ class EmailProviderService extends Component
             ->one();
 
         return new ProviderModel($result);
-    }
-
-    public function saveProvider(ProviderModel $model, bool $runValidation = true)
-    {
-        if ($model->id) {
-            $record = ProviderRecord::findOne($model->id);
-
-            if (!$record) {
-                throw new Exception(Craft::t('emailvalidator',
-                    'No provider exists with the ID "{id}"',
-                    ['id' => $model->id]
-                ));
-            }
-        } else {
-            $record = new ProviderRecord();
-        }
-
-        if ($runValidation && !$model->validate()) {
-            Craft::info('Provider not saved due to a validation error.', __METHOD__);
-            return false;
-        }
-
-        $record->type   = $model->type;
-        $record->domain = $model->domain;
-
-        // Save it
-        $record->save(false);
-
-        // Now that we have a record id, save it on the model
-        $model->id = $record->id;
-
-        return true;
     }
 
     // Private Methods